backend/pyproject.toml
[project]
name = "meeting-ai-backend"
version = "1.0.0"
description = "AI Meeting & Notes Assistant - Backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "redis>=5.0.0",
    "celery>=5.4.0",
    "httpx>=0.27.0",
    "aiofiles>=23.2.0",
    "pydub>=0.25.1",
    "openai>=1.30.0",
    "anthropic>=0.25.0",
    "google-generativeai>=0.5.0",
    "mistralai>=0.4.0",
    "botbuilder-core>=4.16.0",
    "botbuilder-schema>=4.16.0",
    "msal>=1.28.0",
    "reportlab>=4.1.0",
    "markdown>=3.6",
    "websockets>=12.0",
    "python-dateutil>=2.9.0",
    "orjson>=3.10.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
